day5 solution

+++ src/template.rs
@@ -1,6 +1,6 @@
 use crate::utils::read_input;
 
-const DAY: usize = 0;
+const DAY: u8 = 0;
 
 pub fn camp_cleanup() {
     let input = read_input(DAY, true);
